The Pico W now serves a basic HTML page.
This has not been so simple. The only way I could find to make picoserve
and embassy compatible was to get a local copy of picoserve and change
its embassy dependencies to use git as their source. Otherwise there are
conflicts about embassy-timer-driver versions. Using picoserve also
required changing my Rust "channel" to "nightly". A bunch of stuff I
am not keen on, but necessary to progress rather than get bogged down in
build system meta.

Meanwhile also I've changed this to use a static IP at 192.168.3.14 for
now. For currently unknownr reasons the loop waiting for the DHCP lease
to work isn't exiting. I have rebooted the DHCP server and that didn't
help and given the WiFi on the Pico W is working I don't really know
whats up.