IT WORKS! OK, for small values of works, and the code is a mess.
I just got a SDC40 CO2 PPM sensor in the post and I wanted to give that
some basic verification testing... so I hooked it up to the Pico W and
and found a libscd Rust library. I got this successfully reading data,
but then I wanted to get that data available to serve via a HTTP
request. This took a long time to work out... days 2 to 5 of my Rust
journey thus far.

This checkin is lilted with all the detritus of the learning process,
lots of commented out failed attempt code... next job is to clean up the
state of it before further implementation.